pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "AI agents for autonomous software project management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ai", "agents", "langgraph", "github", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",  # For Cerebras z.ai (OpenAI-compatible API)
    "langfuse>=2.0.0",
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyGithub>=2.0.0",
    "PyYAML>=6.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    # Note: Codex CLI is installed via npm (npm install -g @openai/codex), not pip
    # Google APIs for Chat API notifications
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
bedrock = [
    "boto3>=1.34.0",
    "langchain-aws>=0.2.0",
]

[project.scripts]
ai-agents = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "cli", "config"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",  # Line length handled by black
    "E741",  # Ambiguous variable name (l in list comprehensions is fine)
    "E402",  # Module level import not at top (needed for deprecation warnings)
    "UP042",  # Allow (str, Enum) pattern - StrEnum requires Python 3.11+ migration
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
